+++ angr/state_plugins/nvariables.py
